chore: forge fmt

+++ test/invariants/utils/StdAsserts.sol
@@ -5,24 +5,24 @@ import {PropertiesAsserts} from "./PropertiesAsserts.sol";
 import {stdMath} from "forge-std/StdMath.sol";
 
 /// @notice Standardized assertions for use in Invariant tests, inherits PropertiesAsserts
-/// @dev Adapted from forge to work with echidna & medusa 
+/// @dev Adapted from forge to work with echidna & medusa
 abstract contract StdAsserts is PropertiesAsserts {
     event log(string);
     event logs(bytes);
 
     event log_address(address);
     event log_bytes32(bytes32);
-    event log_int(int);
-    event log_uint(uint);
+    event log_int(int256);
+    event log_uint(uint256);
     event log_bytes(bytes);
     event log_string(string);
 
     event log_named_address(string key, address val);
     event log_named_bytes32(string key, bytes32 val);
-    event log_named_decimal_int(string key, int val, uint decimals);
-    event log_named_decimal_uint(string key, uint val, uint decimals);
-    event log_named_int(string key, int val);
-    event log_named_uint(string key, uint val);
+    event log_named_decimal_int(string key, int256 val, uint256 decimals);
+    event log_named_decimal_uint(string key, uint256 val, uint256 decimals);
+    event log_named_int(string key, int256 val);
+    event log_named_uint(string key, uint256 val);
     event log_named_bytes(string key, bytes val);
     event log_named_string(string key, string val);
     event log_array(uint256[] val);
@@ -63,13 +63,10 @@ abstract contract StdAsserts is PropertiesAsserts {
         assertTrue(!data, err);
     }
 
-    function checkEq0(
-        bytes memory a,
-        bytes memory b
-    ) internal pure returns (bool ok) {
+    function checkEq0(bytes memory a, bytes memory b) internal pure returns (bool ok) {
         ok = true;
         if (a.length == b.length) {
-            for (uint i = 0; i < a.length; i++) {
+            for (uint256 i = 0; i < a.length; i++) {
                 if (a[i] != b[i]) {
                     ok = false;
                 }
@@ -88,11 +85,7 @@ abstract contract StdAsserts is PropertiesAsserts {
         }
     }
 
-    function assertEq0(
-        bytes memory a,
-        bytes memory b,
-        string memory err
-    ) internal {
+    function assertEq0(bytes memory a, bytes memory b, string memory err) internal {
         if (!checkEq0(a, b)) {
             emit log_named_string("Error", err);
             assertEq0(a, b);
@@ -119,11 +112,7 @@ abstract contract StdAsserts is PropertiesAsserts {
         assertEq0(a, b);
     }
 
-    function assertEq(
-        bytes memory a,
-        bytes memory b,
-        string memory err
-    ) internal virtual {
+    function assertEq(bytes memory a, bytes memory b, string memory err) internal virtual {
         assertEq0(a, b, err);
     }
 
@@ -154,33 +143,21 @@ abstract contract StdAsserts is PropertiesAsserts {
         }
     }
 
-    function assertEq(
-        uint256[] memory a,
-        uint256[] memory b,
-        string memory err
-    ) internal virtual {
+    function assertEq(uint256[] memory a, uint256[] memory b, string memory err) internal virtual {
         if (keccak256(abi.encode(a)) != keccak256(abi.encode(b))) {
             emit log_named_string("Error", err);
             assertEq(a, b);
         }
     }
 
-    function assertEq(
-        int256[] memory a,
-        int256[] memory b,
-        string memory err
-    ) internal virtual {
+    function assertEq(int256[] memory a, int256[] memory b, string memory err) internal virtual {
         if (keccak256(abi.encode(a)) != keccak256(abi.encode(b))) {
             emit log_named_string("Error", err);
             assertEq(a, b);
         }
     }
 
-    function assertEq(
-        address[] memory a,
-        address[] memory b,
-        string memory err
-    ) internal virtual {
+    function assertEq(address[] memory a, address[] memory b, string memory err) internal virtual {
         if (keccak256(abi.encode(a)) != keccak256(abi.encode(b))) {
             emit log_named_string("Error", err);
             assertEq(a, b);
@@ -192,11 +169,7 @@ abstract contract StdAsserts is PropertiesAsserts {
         assertEq(uint256(a), uint256(b));
     }
 
-    function assertApproxEqAbs(
-        uint256 a,
-        uint256 b,
-        uint256 maxDelta
-    ) internal virtual {
+    function assertApproxEqAbs(uint256 a, uint256 b, uint256 maxDelta) internal virtual {
         uint256 delta = stdMath.delta(a, b);
 
         if (delta > maxDelta) {
@@ -209,12 +182,7 @@ abstract contract StdAsserts is PropertiesAsserts {
         }
     }
 
-    function assertApproxEqAbs(
-        uint256 a,
-        uint256 b,
-        uint256 maxDelta,
-        string memory err
-    ) internal virtual {
+    function assertApproxEqAbs(uint256 a, uint256 b, uint256 maxDelta, string memory err) internal virtual {
         uint256 delta = stdMath.delta(a, b);
 
         if (delta > maxDelta) {
@@ -223,12 +191,7 @@ abstract contract StdAsserts is PropertiesAsserts {
         }
     }
 
-    function assertApproxEqAbsDecimal(
-        uint256 a,
-        uint256 b,
-        uint256 maxDelta,
-        uint256 decimals
-    ) internal virtual {
+    function assertApproxEqAbsDecimal(uint256 a, uint256 b, uint256 maxDelta, uint256 decimals) internal virtual {
         uint256 delta = stdMath.delta(a, b);
 
         if (delta > maxDelta) {
@@ -256,11 +219,7 @@ abstract contract StdAsserts is PropertiesAsserts {
         }
     }
 
-    function assertApproxEqAbs(
-        int256 a,
-        int256 b,
-        uint256 maxDelta
-    ) internal virtual {
+    function assertApproxEqAbs(int256 a, int256 b, uint256 maxDelta) internal virtual {
         uint256 delta = stdMath.delta(a, b);
 
         if (delta > maxDelta) {
@@ -273,12 +232,7 @@ abstract contract StdAsserts is PropertiesAsserts {
         }
     }
 
-    function assertApproxEqAbs(
-        int256 a,
-        int256 b,
-        uint256 maxDelta,
-        string memory err
-    ) internal virtual {
+    function assertApproxEqAbs(int256 a, int256 b, uint256 maxDelta, string memory err) internal virtual {
         uint256 delta = stdMath.delta(a, b);
 
         if (delta > maxDelta) {
@@ -287,12 +241,7 @@ abstract contract StdAsserts is PropertiesAsserts {
         }
     }
 
-    function assertApproxEqAbsDecimal(
-        int256 a,
-        int256 b,
-        uint256 maxDelta,
-        uint256 decimals
-    ) internal virtual {
+    function assertApproxEqAbsDecimal(int256 a, int256 b, uint256 maxDelta, uint256 decimals) internal virtual {
         uint256 delta = stdMath.delta(a, b);
 
         if (delta > maxDelta) {
@@ -333,11 +282,7 @@ abstract contract StdAsserts is PropertiesAsserts {
             emit log("Error: a ~= b not satisfied [uint]");
             emit log_named_uint("        Left", a);
             emit log_named_uint("       Right", b);
-            emit log_named_decimal_uint(
-                " Max % Delta",
-                maxPercentDelta * 100,
-                18
-            );
+            emit log_named_decimal_uint(" Max % Delta", maxPercentDelta * 100, 18);
             emit log_named_decimal_uint("     % Delta", percentDelta * 100, 18);
             fail();
         }
@@ -373,11 +318,7 @@ abstract contract StdAsserts is PropertiesAsserts {
             emit log("Error: a ~= b not satisfied [uint]");
             emit log_named_decimal_uint("        Left", a, decimals);
             emit log_named_decimal_uint("       Right", b, decimals);
-            emit log_named_decimal_uint(
-                " Max % Delta",
-                maxPercentDelta * 100,
-                18
-            );
+            emit log_named_decimal_uint(" Max % Delta", maxPercentDelta * 100, 18);
             emit log_named_decimal_uint("     % Delta", percentDelta * 100, 18);
             fail();
         }
@@ -400,11 +341,7 @@ abstract contract StdAsserts is PropertiesAsserts {
         }
     }
 
-    function assertApproxEqRel(
-        int256 a,
-        int256 b,
-        uint256 maxPercentDelta
-    ) internal virtual {
+    function assertApproxEqRel(int256 a, int256 b, uint256 maxPercentDelta) internal virtual {
         if (b == 0) return assertEq(a, b); // If the left is 0, right must be too.
 
         uint256 percentDelta = stdMath.percentDelta(a, b);
@@ -413,22 +350,13 @@ abstract contract StdAsserts is PropertiesAsserts {
             emit log("Error: a ~= b not satisfied [int]");
             emit log_named_int("        Left", a);
             emit log_named_int("       Right", b);
-            emit log_named_decimal_uint(
-                " Max % Delta",
-                maxPercentDelta * 100,
-                18
-            );
+            emit log_named_decimal_uint(" Max % Delta", maxPercentDelta * 100, 18);
             emit log_named_decimal_uint("     % Delta", percentDelta * 100, 18);
             fail();
         }
     }
 
-    function assertApproxEqRel(
-        int256 a,
-        int256 b,
-        uint256 maxPercentDelta,
-        string memory err
-    ) internal virtual {
+    function assertApproxEqRel(int256 a, int256 b, uint256 maxPercentDelta, string memory err) internal virtual {
         if (b == 0) return assertEq(a, b, err); // If the left is 0, right must be too.
 
         uint256 percentDelta = stdMath.percentDelta(a, b);
@@ -439,12 +367,7 @@ abstract contract StdAsserts is PropertiesAsserts {
         }
     }
 
-    function assertApproxEqRelDecimal(
-        int256 a,
-        int256 b,
-        uint256 maxPercentDelta,
-        uint256 decimals
-    ) internal virtual {
+    function assertApproxEqRelDecimal(int256 a, int256 b, uint256 maxPercentDelta, uint256 decimals) internal virtual {
         if (b == 0) return assertEq(a, b); // If the left is 0, right must be too.
 
         uint256 percentDelta = stdMath.percentDelta(a, b);
@@ -453,11 +376,7 @@ abstract contract StdAsserts is PropertiesAsserts {
             emit log("Error: a ~= b not satisfied [int]");
             emit log_named_decimal_int("        Left", a, decimals);
             emit log_named_decimal_int("       Right", b, decimals);
-            emit log_named_decimal_uint(
-                " Max % Delta",
-                maxPercentDelta * 100,
-                18
-            );
+            emit log_named_decimal_uint(" Max % Delta", maxPercentDelta * 100, 18);
             emit log_named_decimal_uint("     % Delta", percentDelta * 100, 18);
             fail();
         }
@@ -480,11 +399,7 @@ abstract contract StdAsserts is PropertiesAsserts {
         }
     }
 
-    function assertEqCall(
-        address target,
-        bytes memory callDataA,
-        bytes memory callDataB
-    ) internal virtual {
+    function assertEqCall(address target, bytes memory callDataA, bytes memory callDataB) internal virtual {
         assertEqCall(target, callDataA, target, callDataB, true);
     }
 
@@ -513,27 +428,15 @@ abstract contract StdAsserts is PropertiesAsserts {
         bytes memory callDataB,
         bool strictRevertData
     ) internal virtual {
-        (bool successA, bytes memory returnDataA) = address(targetA).call(
-            callDataA
-        );
-        (bool successB, bytes memory returnDataB) = address(targetB).call(
-            callDataB
-        );
+        (bool successA, bytes memory returnDataA) = address(targetA).call(callDataA);
+        (bool successB, bytes memory returnDataB) = address(targetB).call(callDataB);
 
         if (successA && successB) {
-            assertEq(
-                returnDataA,
-                returnDataB,
-                "Call return data does not match"
-            );
+            assertEq(returnDataA, returnDataB, "Call return data does not match");
         }
 
         if (!successA && !successB && strictRevertData) {
-            assertEq(
-                returnDataA,
-                returnDataB,
-                "Call revert data does not match"
-            );
+            assertEq(returnDataA, returnDataB, "Call revert data does not match");
         }
 
         if (!successA && successB) {
